requirements.txt
nflreadpy
pandas
sqlalchemy
psycopg2-binary
jupyter
